==> 0_generate_change_direct.ipynb <==
import pandas as pd

folder = 'd:/git/project1/generated'

prices = pd.read_csv(f'{folder}/bts_price_changes.csv')

prices

res = pd.DataFrame()

res['direct'] = prices.apply(lambda x: 1 if (x.Close - x.Open) > 0 else 0, axis=1)

res['timestamp'] = prices['timestamp']

res

res.to_csv(f'{folder}/btc_daily_change_direct.csv')
