==> 1_bts_price_changes_generate.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:\git\project1'

btc_usd_df = pd.read_csv(f'{folder}/BTC-USD.csv', delimiter=',')
btc_usd_df

def get_change(df, period):
    change = []
    #print(df.shape[0])
    for i in range(df.shape[0]):
        if (i >= period):
            change.append(df.Open[i] - df.Open[i-period])
        else:
            change.append(0)
    return change

period = 2
btc_usd_df[f'change_{period}_day'] = get_change(btc_usd_df, period)
btc_usd_df

import time, datetime
def get_timestamp(date):
    return int(time.mktime(datetime.datetime.strptime(date, "%Y-%m-%d").timetuple()))

btc_usd_df['timestamp'] = btc_usd_df['Date'].apply(get_timestamp)
btc_usd_df

nan_rows = btc_usd_df[btc_usd_df['Open'].isnull()]
nan_rows

btc_usd_df = btc_usd_df.dropna()
btc_usd_df.corr()['change_2_day'].sort_values()

df.to_csv(f'{folder}/generated/bts_price_changes.csv')
