==> 2_generate_words_list.ipynb <==
#nltk.download('stopwords')

import nltk
from nltk.corpus import stopwords
from sklearn.metrics import mean_squared_error
import re
from nltk.corpus import stopwords
import pymorphy2
import pandas as pd

def review_to_wordlist(review):
    review_text = re.sub("[^а-яА-Яa-zA-Z]"," ", review)
    words = review_text.lower().split()
    words = [w for w in words if not w in stops]
    words = [morph.parse(w)[0].normal_form for w in words ]
    return(words)

morph = pymorphy2.MorphAnalyzer()
stops = set(stopwords.words('english'))

df = pd.read_csv("/content/datetime_news.csv")
df

%%time
df['title'] = df['title'].apply(review_to_wordlist)

df.to_csv("/content/date_time_words.csv")
