==> 3_generate_news_clear.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:/git/project1/generated'

df = pd.read_csv(f'{folder}/Headline_Crypto.csv')
df

import re
def is_latin(cell):
    return bool(not re.search(r'[^0-9a-zA-Z~`!@#$%^&*()_+{}\[\];:\'",<.>/?\\| -_=+]', cell))

df['latin'] = df['Headline'].apply(is_latin)
df_clear = df[~(df['latin'] == False)]
df_clear

df_clear.to_csv(f'{folder}/clear_news.csv')
