==> 4_generate_news_small.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:/git/project1/generated'

df_clear = pd.read_csv(f'{folder}/news_clear.csv')

def line_lenght(cell, min, max):
    return bool(cell.count(" ")+1 >= min) and (cell.count(" ")+1 <= max)

df_clear['norm_len'] = df_clear['title'].apply(line_lenght, min=3, max=20)
df_clear[df_clear['norm_len'] == True]

df_small = df_clear[~(df_clear['norm_len'] == False)]
df_small

df_small.to_csv('/content/news_small.csv')
