==> 6_generate_classification_from_tokens.ipynb <==
import torch
import pandas as pd
import numpy as np
import pytorch_transformers

folder = 'd:\git\project1\generated'

model_class = pytorch_transformers.BertModel
tokenizer_class = pytorch_transformers.BertTokenizer
pretrained_weights = 'bert-base-uncased'

model = model_class.from_pretrained(pretrained_weights)

tokens_padded = pd.read_csv(f'{folder}/tokens_padded.csv')
tokens_padded

padded = np.array(tokens_padded)
attention_mask = np.where(padded != 0, 1, 0)
input_ids = torch.tensor(padded)  
attention_mask = torch.tensor(attention_mask)

input_ids

rows_count = input_ids.shape[0]
chank_size = 1000
for chank_num in range (403, (int)(rows_count/1000)+1):
    print(chank_num)
    with torch.no_grad():
        start_index = chank_num*chank_size
        end_index = (chank_num+1)*chank_size-1
        if (end_index > rows_count):
            end_index=rows_count-1
        print(start_index)
        print(end_index)
        input_ids[start_index:end_index, :]
        attention_mask[start_index:end_index, :]
        last_hidden_states = model(input_ids[start_index:end_index, :], attention_mask=attention_mask[start_index:end_index, :])
        classification = pd.DataFrame(last_hidden_states[0][:,0,:].numpy())
        classification.to_csv(f'{folder}\classification_{start_index}_{end_index}.csv')

rows_count = input_ids.shape[0]
chank_size = 1000
for chank_num in range (208, 251):
    print(chank_num)
    with torch.no_grad():
        start_index = chank_num*chank_size
        end_index = (chank_num+1)*chank_size-1
        if (end_index > rows_count):
            end_index=rows_count-1
        print(start_index)
        print(end_index)
        input_ids[start_index:end_index, :]
        attention_mask[start_index:end_index, :]
        last_hidden_states = model(input_ids[start_index:end_index, :], attention_mask=attention_mask[start_index:end_index, :])
        classification = pd.DataFrame(last_hidden_states[0][:,0,:].numpy())
        classification.to_csv(f'{folder}\classification_{start_index}_{end_index}.csv')

last_hidden_states[0][:,0,:].numpy()

classification = pd.DataFrame(last_hidden_states[0][:,0,:].numpy())
classification

classification.to_csv(f'{folder}\classification_array_0_1000.csv')
