==> 7_merge_news_cls_direct.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:/git/project1/generated'

#btc_direct_table = pd.read_csv(f'{folder}/btc_daily_change_direct.csv')
btc_direct_table = pd.read_csv(f'{folder}/bts_price_changes.csv')
btc_direct_table

date_time_title_short = pd.read_csv(f'{folder}/date_time_title_short.csv')

btc_timestamps = btc_direct_table['timestamp']
#btc_direct =  btc_direct_table['direct']
btc_direct =  btc_direct_table['direct_30']
news_timestamps_all = date_time_title_short['date']

def get_directs(news_timestamps):
    directs = []
    for news_timestamp in news_timestamps:
        #print(ts_news)
        finded = False
        for i in range(btc_direct.shape[0]):
            if news_timestamp == btc_timestamps[i]:
                directs.append(btc_direct[i])
                finded = True
                break;
        if (not finded):
            print('not finded')
            directs.append(0)    
    return directs

def get_union_table(num):
    table = pd.read_csv(f'{folder}/classification_{num*1000}_{(num+1)*1000-1}.csv')
    table['timestamp'] = np.array(news_timestamps_all[num*1000:(num+1)*1000-1])
    table['direct'] = get_directs(table['timestamp'])
    return table
    

%%time
res_table = pd.DataFrame(None)
res_table = res_table.append(get_union_table(0), sort=False, ignore_index=True)
res_table

%%time
start = 0
for j in range(2):#max=60
    res_table = pd.DataFrame(None)#clear_table
    for i in range(10):
        num = i+j*10+start
        print(num)
        res_table = res_table.append(get_union_table(num), sort=False, ignore_index=True)
    res_table.to_csv(f'{folder}\cls_direct30_{num}.csv', index=False)

btc_direct_table.loc[btc_direct_table.timestamp == 1550793600, :]

data = pd.read_csv(f'{folder}/set_133.csv')

for i in range(14, 40):
    print (i)
    data_add = pd.read_csv(f'{folder}/set_{i}3.csv')
    data = data.append(data_add, sort=False, ignore_index=True)

data.to_csv(f'{folder}/set_393.csv')
