==> 8_bts_usd_fit.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:/git/project1/generated'

df = pd.read_csv(f'{folder}/bts_price_changes.csv')
del df['Unnamed: 0']
df = df.drop(index=[2400])

df

period = 180
df[f'change_percent_{period}'] = df.apply(lambda x: x[f'change_{period}_day']/x['Close'], axis=1)
target = f'direct_{period}'
df[['timestamp', 'Open', 'High', 'Low', 'Close', 'Volume', target]].corr()[target].sort_values()

x_train = np.array(df.loc[:1800,['timestamp', 'Open', 'High', 'Low', 'Close', 'Volume']])
x_train

x_train = np.array(df.loc[:1800,[target]])
x_train

import tensorflow
from keras import models
from keras import layers
from keras import regularizers
import matplotlib.pyplot as plt

from tensorflow.keras.callbacks import ModelCheckpoint, TensorBoard, ReduceLROnPlateau
weights_file = f'{folder}/weights3.h5'
callback = ModelCheckpoint(weights_file, monitor='val_binary_accuracy', mode='max', save_best_only=True)
tensorboard_cbk = TensorBoard(log_dir=f'{folder}/logs')
reduce = ReduceLROnPlateau(monitor="val_loss", patience=5, min_lr=0.000001, verbose=1)

model = models.Sequential()

model.add(layers.Dense(384, input_shape=(768,), activity_regularizer=regularizers.l2(0.02), kernel_initializer="he_uniform"))
model.add(layers.BatchNormalization())
model.add(layers.ReLU())

model.add(layers.Dropout(0.5))

model.add(layers.Dense(192, activity_regularizer=regularizers.l2(0.02), kernel_initializer="he_uniform"))
model.add(layers.BatchNormalization())
model.add(layers.ReLU())

model.add(layers.Dense(1, activation='sigmoid'))

#model.compile(optimizer='rmsprop', loss='binary_crossentropy', metrics=['binary_accuracy'])
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['binary_accuracy'])

json_file = f'{folder}/model_btc.json'
model_json = model.to_json()

with open(json_file, 'w') as file:
    file.write(model_json)

history = model.fit(x_train, y_train, epochs=5, batch_size=2000, validation_split=0.3, callbacks=[callback, tensorboard_cbk, reduce])

start = 0
loss_values = history.history['loss'][start:]
val_loss_values = history.history['val_loss'][start:]
epochs = range(start, len(loss_values) + start)
plt.plot(epochs, loss_values, 'g', label='Training loss')
plt.plot(epochs, val_loss_values, 'b', label='Validation loss')
plt.title('Training and validation loss')
plt.xlabel('Epochs')
plt.ylabel('Loss')
plt.legend()

plt.clf()
acc_values = history.history['binary_accuracy']
val_acc_values = history.history['val_binary_accuracy']
epochs = range(1, len(acc_values) + 1)
plt.plot(epochs, acc_values, 'g', label='Training acc')
plt.plot(epochs, val_acc_values, 'b', label='Validation acc')
plt.title('Training and validation binary_accuracy')
plt.xlabel('Epochs')
plt.ylabel('binary_accuracy')
plt.legend()
plt.show()
