==> 8.1_bts_usd_fit.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:/git/project1/generated'

df = pd.read_csv(f'{folder}/bts_price_changes.csv')
del df['Unnamed: 0']
df = df.drop(index=[2400])

period = 180
df[f'change_percent_{period}'] = df.apply(lambda x: x[f'change_{period}_day']/x['Close'], axis=1)
target = f'change_percent_{period}'
df[['timestamp', 'Open', 'High', 'Low', 'Close', 'Volume', target]].corr()[target].sort_values()

df.to_csv(f'{folder}/bts_price_changes.csv')

x_train = np.array(df.loc[:1800,['timestamp', 'Open', 'High', 'Low', 'Close', 'Volume']])
x_train

target = f'direct_{period}'
y_train = np.array(df.loc[:1800,[target]])
y_train

x_test = np.array(df.loc[1801:,['timestamp', 'Open', 'High', 'Low', 'Close', 'Volume']])
x_test

y_test = np.array(df.loc[1801:,[target]])
y_test

from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import cross_val_score

%%time
parameters = {'C': np.linspace(0.0001, 100, 20)}
grid_search = GridSearchCV(LogisticRegression(), parameters)
grid_search.fit(x_train, y_train.ravel())

print('best parameters: ', grid_search.best_params_)
print('best scrores: ', grid_search.best_score_)

%%time
clf = LogisticRegression(C=0.0001)
clf.fit(x_train, y_train.ravel())

LinearRegression().get_params().keys()

clf.score(x_test, y_test)

from sklearn.dummy import DummyClassifier
clf_dummy = DummyClassifier()

scores = cross_val_score(clf_dummy, x_test, y_test)
print("Dummy classifier score: %0.3f (+/- %0.2f)" % (scores.mean(), scores.std() * 2))

from sklearn.ensemble import RandomForestRegressor


%%time
rf = RandomForestRegressor(n_estimators = 1000)
rf.fit(x_train, y_train.ravel())

rf_pred = rf.predict(x_test)

threshold = 0.5
preds = [1 if val > threshold else 0 for val in rf_pred]

from sklearn.metrics import mean_squared_error
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score

#print(roc_auc_score(preds, y_test))
print(confusion_matrix(preds, y_test))
print(accuracy_score(preds, y_test))
