==> 9_Headline_Crypto_data_prepare.ipynb <==
import pandas as pd
import numpy as np
folder = 'd:\git\project1\generated'

df = pd.read_csv(f'{folder}/Headline_Crypto.csv')
df

def get_data(x):
    try:
        res = int(time.mktime(datetime.datetime.strptime(x.split('T')[0], "%Y-%m-%d").timetuple()))
    except ValueError:
        res = None
    return res

import time, datetime
df['date'] = df['Date'].apply(lambda x: get_data(x))
#df['time'] = time_to_int(datetime_string.split()[1])

df = df.dropna()
df

df.groupby('date').nunique()
