==> 10_btc_prepare.ipynb <==
import pandas as pd
import numpy as np
dir = 'd:\git\project1\data'

btc_usd_df = pd.read_csv(f'{dir}/BTC_USD_2013-10-01_2021-05-06-CoinDesk.csv', delimiter=',')
btc_usd_df

import time, datetime
def get_timestamp(date):
    return int(time.mktime(datetime.datetime.strptime(date, "%Y-%m-%d").timetuple()))

btc_usd_df['timestamp'] = btc_usd_df['Date'].apply(get_timestamp)
btc_usd_df

del btc_usd_df['Currency']
del btc_usd_df['Date']

btc_usd_df = btc_usd_df.dropna()
btc_usd_df

def get_period_change(df, period):
    change = []
    #print(df.shape[0])
    for i in range(df.shape[0]):
        if (i >= period):
            change.append(df.loc[i, 'Closing Price (USD)'] - df.loc[i-period, 'Closing Price (USD)'])
        else:
            change.append(0)
    return change

def get_percent_change(df, period):
    change = []
    #print(df.shape[0])
    for i in range(df.shape[0]):
        if (i >= period):
            change.append((df.loc[i, 'Closing Price (USD)'] - df.loc[i-period, 'Closing Price (USD)'])/df.loc[i, 'Closing Price (USD)'])
        else:
            change.append(0)
    return change

def get_direct_change(df, period):
    change = []
    #print(df.shape[0])
    for i in range(df.shape[0]):
        if (i >= period):
            change.append(1 if (df.loc[i, 'Closing Price (USD)'] - df.loc[i-period, 'Closing Price (USD)']) > 0 else 0)
        else:
            change.append(0)
    return change

#period = 1
#btc_usd_df[f'change_{period}_day'] = get_period_change(btc_usd_df, period)
#btc_usd_df

#period = 1
#btc_usd_df[f'change_{period}_percent'] = get_percent_change(btc_usd_df, period)
#btc_usd_df

period = 90
btc_usd_df[f'change_{period}_direct'] = get_direct_change(btc_usd_df, period)
btc_usd_df

btc_usd_df.to_csv(f'{dir}/prepared_btc.csv', index=False)

btc_usd_df.corr()['change_1_direct'].sort_values()
