==> 10_tweets_prepare.ipynb <==
import pandas as pd
import numpy as np
dir = 'd:\git\project1\data'

#%%time
#cols = ['timestamp','retweets','text']
#df = pd.read_csv(f'{dir}/tweets.csv', delimiter=';', usecols=cols, skiprows=500000, nrows=500000, names=['id','user','fullname','url','timestamp','replies','likes','retweets','text'], dtype={'fullname': str} )
#df = pd.read_csv(f'{dir}/tweets.csv', delimiter=';', usecols=cols)
#df

#df_with_retweet = df[df['retweets'] > 1]
#df_with_retweet

#df_with_retweet.to_csv(f'{dir}/tweets_with_retweet.csv', index=False)

df = pd.read_csv(f'{dir}/tweets_with_retweet.csv')
df

import time, datetime
df['date'] = df['timestamp'].apply(lambda x:int(time.mktime(datetime.datetime.strptime(x.split()[0], "%Y-%m-%d").timetuple())))

def time_to_int(time_str):
    time_str = time_str.split('+')[0]
    hh, mm , ss = map(int, time_str.split(':'))
    return ss + 60*(mm + 60*hh)
df['time'] = df['timestamp'].apply(lambda x:time_to_int(x.split()[1]))

#df['latin'] = df['text'].apply(lambda x: bool(re.search(r'\b\w+\w+\w+\b', x, flags=re.ASCII)))
#df_clear = df[~(df['latin'] == False)]
#df_clear

del df['timestamp']

df

cols = ['timestamp', 'change_1_direct', 'change_3_direct', 'change_7_direct', 'change_30_direct', 'change_90_direct']
btc_usd_df = pd.read_csv(f'{dir}/prepared_btc.csv', usecols=cols)
btc_usd_df

union = df.merge(btc_usd_df, how='left', left_on='date', right_on='timestamp')

union.dropna(inplace=True)

union

union.to_csv(f'{dir}/tweets_change.csv', index=False)

union[target].value_counts()

union.corr()[target].sort_values()

#df_day = df[df['date'] == 1336946400]
#df_day_max = df_day[df_day['retweets'] == df_day['retweets'].max()]
#df_day_max

text = df_day_max['text'].iloc[0]
#text = '@bitcoin hacked? @but it was going @to replace gold...'
text = re.sub(r'@\w+', '', text, count=0)
print (text)

df_clear.groupby('date').nunique()

df.groupby('date').nunique()

df.corr()['retweets'].sort_values()
