==> 11_pr3_tweets_tokens.ipynb <==
import pandas as pd
import numpy as np
dir = 'd:\git\project1\data'

import torch
import pytorch_transformers
model_class = pytorch_transformers.BertModel
tokenizer_class = pytorch_transformers.BertTokenizer
pretrained_weights = 'bert-base-uncased'

tokenizer = tokenizer_class.from_pretrained(pretrained_weights)
model = model_class.from_pretrained(pretrained_weights)

df = pd.read_csv(f'{dir}/tweets.csv', delimiter=';', nrows=100000)

df['tw'] = df['retweets'].apply(lambda x: 1 if x>0 else 0)

df.corr()['tw']

del df['replies']
del df['likes']
del df['user']
del df['id']
del df['fullname']
del df['url']
del df['timestamp']
df

%%time
tokenized = df['text'].apply((lambda x: tokenizer.encode(x, add_special_tokens=True)))
tokenized

max_len = 0
for i in tokenized:
    if len(i) > max_len:
        max_len = len(i)
print(max_len)

tokenized.apply(lambda x: len(x)).sort_values()

tokens = pd.DataFrame()
tokens['retweets'] = df['retweets'].apply(lambda x: 1 if x>0 else 0)
max = 50
tokens['raw'] = tokenized.apply(lambda x: x if (len(x) < max) else None)
#tokens = tokens.dropna()
#tokens = tokenized.apply(token_lenght, max=30)

#tokens['raw'] = tokenized
tokens

#tokens['retweets'].groupby('retweets').nunique()
tokens.groupby('retweets').count()

%matplotlib
tokens['retweets'].plot.hist(bins=1000)

%%time
padded = np.array([i + [0]*(max-len(i)) if i!=None else [None]*(max) for i in tokens['raw']])
padded

df_padded = pd.DataFrame(padded)
df_padded

%%time
union = pd.concat([tokens['retweets'], df_padded], axis=1)
union.dropna(inplace=True)
union

%%time
union.to_csv(f'{dir}/retweets_padded1.csv', index=False)
