==> 12_tokens_to_cls.ipynb <==
import torch
import pandas as pd
import numpy as np
import pytorch_transformers

folder = 'd:\git\project1\data'

model_class = pytorch_transformers.BertModel
tokenizer_class = pytorch_transformers.BertTokenizer
pretrained_weights = 'bert-base-uncased'

model = model_class.from_pretrained(pretrained_weights)

%%time
tokens_padded = pd.read_csv(f'{folder}/tweets_change_padded.csv')
tokens_padded

%%time
padded = np.array(tokens_padded.iloc[:,8:])
attention_mask = np.where(padded != 0, 1, 0)
input_ids = torch.tensor(padded)  
attention_mask = torch.tensor(attention_mask)

input_ids

def save_last_hidden(start, count):
    start_index = start*count
    end_index = (start+1)*count
    last_hidden_states = model(input_ids[start_index:end_index, :], attention_mask=attention_mask[start_index:end_index, :])
    last_hidden_states_df = pd.DataFrame(last_hidden_states[0][:,0,:].numpy())
    #print(np.array(tokens_padded.loc[start_index:end_index-1, ['retweets']]))
    #last_hidden_states_df['retweet'] = np.array(tokens_padded.loc[start_index:end_index-1, ['retweets']])
    #last_hidden_states_df['time'] = np.array(tokens_padded.loc[start_index:end_index-1, ['time']])
    #last_hidden_states_df['date'] = np.array(tokens_padded.loc[start_index:end_index-1, ['date']])
    #last_hidden_states_df['change_1_direct'] = np.array(tokens_padded.loc[start_index:end_index-1, ['change_1_direct']])
    #last_hidden_states_df['change_3_direct'] = np.array(tokens_padded.loc[start_index:end_index-1, ['change_3_direct']])
    #last_hidden_states_df['change_7_direct'] = np.array(tokens_padded.loc[start_index:end_index-1, ['change_7_direct']])
    #last_hidden_states_df['change_30_direct'] = np.array(tokens_padded.loc[start_index:end_index-1, ['change_30_direct']])
    #last_hidden_states_df['change_90_direct'] = np.array(tokens_padded.loc[start_index:end_index-1, ['change_90_direct']])
    #print(last_hidden_states_df)
    last_hidden_states_df.to_csv(f'{folder}/last_hs_tweets_ch_{start_index}.csv', index=False)

%%time
rows_count = input_ids.shape[0]
chank_size = 1000
start = 413
for chank_num in range (start, (int)(rows_count/chank_size)+1):
    print(chank_num)
    with torch.no_grad():
        save_last_hidden(chank_num, chank_size)
        

%%time
with torch.no_grad():
    #save_last_hidden(251, 1000)
