==> 13_hs_target_Union.ipynb <==
import pandas as pd
import numpy as np

folder = 'd:\git\project1\data'

%%time
changes = pd.read_csv(f'{folder}/tweets_change_padded.csv')
changes

cols = ['date', 'change_90_direct']
changes.loc[3000:4000, cols]

#last_hidden_states_df = pd.read_csv(f'{folder}/last_hs_tweets_ch1_0.csv')
#last_hidden_states_df

def concat_hs_targets(cols, start_index, count):
    start_index = start*count
    end_index = start_index + count
    last_hidden_states_df = pd.read_csv(f'{folder}/last_hs_tweets_ch_{start_index}.csv')
    df = pd.DataFrame(np.array(changes.loc[start_index:end_index, cols]), columns=cols)
    union = pd.concat([last_hidden_states_df, df], axis=1)
    union.to_csv(f'{folder}/last_hs_tweets_ch_{start_index}.csv', index=False)
    print(start_index)
    print(union['change_90_direct'].value_counts())

for i in range(251):
    concat_hs_targets(cols, i, 1000)

union['change_30_direct'].value_counts()

concat_hs_targets(cols, 0, 10)
